state: improve handling
Always apply rules, before restoring state, otherwise rules callbacks
are not called (which can have a huge impact on the way clients are
configured; e.g. with respect to signal registration).

This also has the additional benefit of alleviating the need for custom
restore/save state handling (e.g in the players code).